WIP on a version of Grid2D.merge that allows extension
I'm shelving this idea for now - it was a cool thought but in practice
it's just too hairy for functionality I don't actually need. Maybe I'll
revisit this at some point, but for now, any usage of merge to "extend"
a grid will be left as an exercise to the user. (i.e. me).